streamlit-chat-app/requirements-streamlit.txt
# Streamlit Chat App with Azure OpenAI
# Core dependencies for the chat application

# Streamlit for web interface
streamlit>=1.28.0

# OpenAI SDK for Azure OpenAI integration
openai>=1.12.0

# Additional dependencies that might be needed
python-dotenv>=1.0.0